Dev sort changes oldest first, fixes separately
There is no big reason for this but I keep wanting it. Somewhat-tangible
bonus is it is consistent with the higher-level ordering where "most
released" changes (e.g. binary release) are reported first, then merged,
then "In progress".

I tried to make no text changes and only do the reordering.