People shorten interview para
So I think interview/trading links are useful in both Media (list of all
new articles in one place) and relevant sections (People stuff and
Markets stuff). We already have it for the State of the Market video in
both Media and relevant Dev places. To reduce the effect of duplication
we can make the copies shorter.