Media add 2 TA posts by Applesaucesome
In 202205 and 202206 I only added these in Markets. In 202207 and 202208
I added these in both Media and Markets. I had a lot of back and forth
on duplicating (or not) posts in Media and a more relevant section. My
current thinking is do it:

- Media seems useful to see *all* article content in one place,
  regardless of its topic. Same for videos.
- It is also nice when media content is mentioned in relevant sections
  like Dev/People/Markets/etc.
- The above use cases justify the repetition in my opinion.
- To reduce the effect of repetition we can make text copy on repeated
  links shorter.

See also 0c2c388 and 4ad2260 for discussion and 1ed2d93 or 9286e8a for
examples of handling this situation. Mind that these commits are in the
`drafts` branch.