journal: Fix anchors
Non-void self-closing tags are not valid in HTML5. `<a>` is non-void and
so `<a ... />` is not valid. In practice this is not a problem right now
but this became a problem in GitHub Releases (which use a different
Markdown renderer) recently after all links gained underline style.

Fixing this now just in case other GitHub renderers change.

- https://stackoverflow.com/questions/3558119/are-non-void-self-closing-tags-valid-in-html5
- https://html.spec.whatwg.org/multipage/parsing.html#parse-error-non-void-html-element-start-tag-with-trailing-solidus